wine_chatbot/requirements.txt
Flask
paddleocr
pillow
numpy
mysql-connector-python
python-dotenv